* commit 'bc7aa6b7f8d56ff20056f2fe43e2c8e1ee2e6ecc':
  Simplify RDX3DPanelNotificationManager
  Dont spam locomotion manager notifications
  Support multiple notifications

+++ ihmc-graphics/src/libgdx/java/us/ihmc/rdx/ui/RDX3DPanelTooltip.java
@@ -34,7 +34,7 @@ public class RDX3DPanelTooltip
     *                                above this one. This is the number of lines of text being
     *                                displayed in that one.
     */
-   public void render(String text, int linesOfSpaceToGiveAbove)
+   public void render(String text, float linesOfSpaceToGiveAbove)
    {
       if (latestInput != null)
       {
